+++ changelog.md
@@ -2,6 +2,7 @@
 - bugfix in `mass_step`, resulting in wrong heights
 - decreased default `samples` for `mass_step` from 20 to 10
 - fixed calibration message
+- made `Worklist.profiles` `@property` and removed option to initialize `Worklist` with different profiles (still possible to add `Worklists` with different profiles)
 
 # 2026-02
 - add debug statements to `read_data` and fix logging level issue.